% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_report.R
\name{arrow_to_mlr3}
\alias{arrow_to_mlr3}
\title{arrow_to_mlr3}
\usage{
arrow_to_mlr3(dataset)
}
\arguments{
\item{dataset}{\link[arrow:Table-class]{arrow::Table} or \link[arrow:Dataset]{arrow::Dataset}}

\item{primary_key}{character name of the column to use as a primary key}
}
\value{
\link[mlr3db:DataBackendDplyr]{mlr3db::DataBackendDplyr}
}
\description{
Converts arrow Table/Dataset to mlr3 Backend
}
