% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance_report.R
\name{attendance_report}
\alias{attendance_report}
\alias{read.attendance_report}
\alias{process.attendance_report}
\title{attendance_report}
\usage{
\method{read}{attendance_report}(
  attendance_report,
  list_no,
  tables = c("orders", "order_detail", "special_activities", "performances",
    "constituencies", "customers", "seats", "attributes"),
  since = Sys.time(),
  until = as.POSIXct(Sys.Date() + 3),
  refresh = TRUE
)

\method{process}{attendance_report}(
  attendance_report,
  sli_statuses = c("Unseated, Unpaid", "Seated, Unpaid", "Seated, Paid",
    "Unseated, Paid", "Upgraded", "Ticketed, Paid"),
  special_activity_statuses = c("Attended", "Accepted", "Tentative")
)
}
\arguments{
\item{attendance_report}{\code{attendance_report} object}

\item{list_no}{\code{integer(1)} list number to load customers from}

\item{tables}{\code{character} vector of tables to load from, as defined by \link[tessilake:tessi_list_tables]{tessilake::tessi_list_tables}}

\item{since}{\code{POSIXct} date from which to query performances}

\item{until}{\code{POSIXct} date up to which to query performances}

\item{refresh}{\code{logical} whether or not to force refresh the data (default is to refresh)}
}
\description{
Sends an email containing a spreadsheet and pdf of attendees
}
\section{Methods (by generic)}{
\itemize{
\item \code{read(attendance_report)}: load data for \code{attendance_report}

\item \code{process(attendance_report)}: process data for \code{attendance_report}

}}
\note{
\itemize{
\item \strong{orders}  : order_no, custom_1
\item \strong{order_detail} : order_no, sli_status_desc, seating_status_desc, perf_no,
perf_dt, perf_desc, recipient_display_name, customer_no, order_ship_method_desc
\item \strong{activities} : customer_no, activity_desc, status_desc
\item \strong{performances} : perf_no, perf_desc, perf_dt
\item \strong{constituencies} : customer_no, constituency_desc
\item \strong{customers} : customer_no, display_name
\item \strong{performance_seating} : perf_no, seat_no, seat_row, seat_num
\item \strong{attributes} : customer_no, keyword_desc, keyword_value
\item \strong{scans} : customer_no, perf_no
}
}
