% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditR.R
\name{new_auditR}
\alias{new_auditR}
\title{auditR}
\usage{
new_auditR(x = new.env(), class = character())
}
\arguments{
\item{x}{an environment}

\item{class}{an optional additional subclass
#'}
}
\description{
Base S3 class for auditR.
}
