% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanAddress.R
\name{cleanAddress}
\alias{cleanAddress}
\title{cleanAddress}
\usage{
cleanAddress(vec)
}
\arguments{
\item{vec}{A character vector of addresses}
}
\value{
A character vector with directions and suffixes mapped to their USPS-approved abbreviated forms,
non-alphanumeric characters mapped to spaces and hyphens (-) removed.
}
\description{
cleanAddress
}
\examples{
vec <- c("25-13 east 25th street","43 1/2 southeast st james place")
cleanAddress(vec)

}
