% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions_model.R
\docType{data}
\name{contributions_model}
\alias{contributions_model}
\title{contributions_model}
\format{
An object of class \code{contributions_model} (inherits from \code{mlr_report}, \code{report}, \code{list}) of length 0.
}
\usage{
contributions_model
}
\description{
mlr3 model for predicting contributions
}
\keyword{datasets}
