% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions_model_report.R
\docType{data}
\name{contributions_model_report}
\alias{contributions_model_report}
\alias{read.contributions_model_report}
\alias{process.contributions_model_report}
\alias{output.contributions_model_report}
\title{contributions_model_report}
\format{
An object of class \code{contributions_model_report} (inherits from \code{report}, \code{list}) of length 0.
}
\usage{
contributions_model_report

\method{read}{contributions_model_report}(report, ...)

\method{process}{contributions_model_report}(report, ...)

\method{output}{contributions_model_report}(
  report,
  subject = paste("Contributions model", Sys.Date()),
  body = paste("Sent by", Sys.info()["nodename"]),
  emails = config::get("tessiflow.email"),
  ...
)
}
\arguments{
\item{report}{\code{contributions_model_report} object}

\item{...}{not used}

\item{subject}{Subject of the email.}

\item{body}{Body of the email as text. If the parameter body refers to an existing file location, the text of the file is parsed as body of the email.}

\item{emails}{email addresses (first will be sender)}
}
\description{
emailed report of predictions from \link{contributions_model}
}
\section{Methods (by generic)}{
\itemize{
\item \code{read(contributions_model_report)}: Read data for contributions_model_report

\item \code{process(contributions_model_report)}: Process data for contributions_model_report

\item \code{output(contributions_model_report)}: Send spreadsheet for contributions_model_report

}}
\keyword{datasets}
