% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossovers.R
\name{crossovers_prod}
\alias{crossovers_prod}
\title{crossovers_prod}
\usage{
crossovers_prod(prod_season_no, min_customers = 10)
}
\arguments{
\item{prod_season_no}{\link{integer} vector of production season numbers to search for crossovers against}

\item{min_customers}{\href{1}{integer} a performance must have at least this number of customers to qualify as
a potential crossover}
}
\description{
Simple analysis of crossover audiences for particular production(s), returns a \link[data.table:data.table]{data.table} of productions,
ranked from highest to lowest \% crossover.
}
