% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_chunk_write}
\alias{dataset_chunk_write}
\title{dataset_chunk_write}
\usage{
dataset_chunk_write(
  dataset,
  partition,
  dataset_name,
  rows = data.table(I = seq_len(nrow(dataset))),
  cols = colnames(dataset),
  rollback = missing(),
  timestamp = missing(),
  ...
)
}
\arguments{
\item{dataset}{\code{data.frameish} dataset to load from}

\item{partition}{\code{character}|\code{integer} identifying the partition the chunk will be saved in}

\item{dataset_name}{\code{character} cache directory where the partition will be saved in}

\item{rows}{\link{data.table} identifying rows of the dataset to load; will be appended to dataset}

\item{cols}{\code{character} columns of the dataset to add to partition}

\item{rollback}{\code{character} passed on to \link{dataset_rollback_event}}

\item{timestamp}{\code{character} passed on to \link{dataset_normalize_timestamps}}

\item{...}{
  Arguments passed on to \code{\link[=dataset_rollback_event]{dataset_rollback_event}}
  \describe{
    \item{\code{event}}{character column name containing a logical feature that indicates events to rollback}
    \item{\code{by}}{character column name to group the table by}
  }}
}
\description{
Write out a chunk of a larger \code{dataset}, using Hadoop partition=\code{partition} nomenclature,
saving it in the cache dir \code{dataset_name}. The chunk is identified by the I column in \code{rows},
which is attached to the columns of the dataset identified by \code{cols}. Features matching
the regular expression in \code{rollback} are rolled back one row, and all timestamps are normalized
by \link{dataset_normalize_timestamps}.
}
