% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_normalize_timestamps}
\alias{dataset_normalize_timestamps}
\title{dataset_normalize_timestamps}
\usage{
dataset_normalize_timestamps(
  dataset,
  timestamp_cols = grep("timestamp", colnames(dataset), value = T, ignore.case = T),
  by = "group_customer_no",
  ...
)
}
\arguments{
\item{dataset}{data.table of data to normalize}

\item{timestamp_cols}{character vector of columns to normalize; defaults to all columns with a name containing the word \code{timestamp}}

\item{by}{character column name to group the table by}

\item{...}{not used}
}
\value{
normalized data.table
}
\description{
Replaces the date-times in \code{columns} with integer offsets from the first \code{timestamp} per group identified by \code{by}.
}
