% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3proba_shim.R
\name{.__TaskSurv__initialize}
\alias{.__TaskSurv__initialize}
\title{.__TaskSurv__initialize}
\usage{
.__TaskSurv__initialize(
  id,
  backend,
  time = "time",
  event = "event",
  time2,
  type = c("right", "left", "interval", "counting", "interval2", "mstate"),
  label = NA_character_
)
}
\arguments{
\item{id}{(\code{character(1)})\cr
Identifier for the new instance.}

\item{backend}{(\link{DataBackend})\cr
Either a \link{DataBackend}, or any object which is convertible to a \link{DataBackend} with \code{as_data_backend()}.
E.g., a \code{data.frame()} will be converted to a \link{DataBackendDataTable}.#' @template param_time}

\item{time}{(\code{character(1)})\cr
Name of the column for event time if data is right censored, otherwise starting time if
interval censored.}

\item{event}{(\code{character(1)})\cr
Name of the column giving the event indicator.
If data is right censored then "0"/\code{FALSE} means alive (no event), "1"/\code{TRUE} means dead
(event). If \code{type} is \code{"interval"} then "0" means right censored, "1" means dead (event),
"2" means left censored, and "3" means interval censored. If \code{type} is \code{"interval2"} then
\code{event} is ignored.}

\item{time2}{(\code{character(1)})\cr
Name of the column for ending time of the interval for interval censored or
counting process data, otherwise ignored.}

\item{type}{(\code{character(1)})\cr
Name of the column giving the type of censoring. Default is 'right' censoring.}

\item{label}{(\code{character(1)})\cr
Label for the new instance.}
}
\description{
Shim for .__TaskSurv__initialize to work with duckdb tables
}
