% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates_report.R
\name{duplicates_report}
\alias{duplicates_report}
\alias{read.duplicates_report}
\alias{output.duplicates_report}
\title{duplicates_report}
\usage{
\method{read}{duplicates_report}(duplicates_report, ...)

\method{output}{duplicates_report}(duplicates_report, routing = NULL, ...)
}
\arguments{
\item{duplicates_report}{duplicates_report object}

\item{...}{
  Arguments passed on to \code{\link[=send_email]{send_email}}
  \describe{
    \item{\code{body}}{html of message or a list containing \link[sendmailR:mime_part]{sendmailR::mime_part} objects.}
  }}

\item{routing}{named list, names are email addresses and values are vectors of
customer numbers. Any duplicate pair that contains a matching customer number is
emailed to the named email address.}
}
\description{
Sends an email containing a spreadsheet of duplicates for review
}
\section{Methods (by generic)}{
\itemize{
\item \code{read(duplicates_report)}: read the duplicates data output by \code{tessistream::duplicates_stream}

\item \code{output(duplicates_report)}: segments the duplicates data based on the map provided
by \code{routing} and emails excel spreadsheets using \link{send_xlsx}

}}
