% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iml.R
\name{iml}
\alias{iml}
\alias{iml_predictor}
\title{interpretable machine learning for mlr_report}
\usage{
iml_predictor(model, data, predict.function = NULL, y = NULL)
}
\arguments{
\item{model}{mlr3 model, pre-trained}

\item{data}{data.table of a test dataset}

\item{predict.function}{function the function to predict newdata. The first argument is the model, the second the newdata.}

\item{y}{character/numeric/factor The target vector or (preferably) the name of the target column in the data argument. Predictor tries to infer the target automatically from the model.}
}
\value{
\link[iml:Predictor]{iml::Predictor}
}
\description{
interpretable machine learning for mlr_report using \link{iml}
}
\section{Functions}{
\itemize{
\item \code{iml_predictor()}: Wrapper around \link[iml:Predictor]{iml::Predictor} to subset the features of data and
provide a \code{predict.function} and \code{y} when the predictor can't identify
them.

}}
