% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{make_unsubscribe_report_fixtures}
\alias{make_unsubscribe_report_fixtures}
\title{make_unsubscribe_report_fixtures}
\usage{
make_unsubscribe_report_fixtures()
}
\description{
Make fixtures for unsubscribe_report testing
}
