% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_report.R
\docType{data}
\name{mlr_report}
\alias{mlr_report}
\alias{read.mlr_report}
\alias{process.mlr_report}
\alias{output.mlr_report}
\alias{write.mlr_report}
\title{mlr_report}
\format{
An object of class \code{mlr_report} (inherits from \code{report}, \code{list}) of length 0.
}
\usage{
mlr_report

\method{read}{mlr_report}(mlr_report)

\method{process}{mlr_report}(mlr_report, train = TRUE, predict = TRUE)

\method{output}{mlr_report}()

\method{write}{mlr_report}(mlr_report, subdir = "models", sync = TRUE)
}
\arguments{
\item{mlr_report}{The report object}

\item{train}{boolean whether to run the training}

\item{predict}{boolean whether to run the predictions}

\item{subdir}{character name of subdir to store data in}

\item{sync}{boolean whether to sync data across storages}
}
\description{
Base class for training/running a machine learning model
}
\section{Methods (by generic)}{
\itemize{
\item \code{read(mlr_report)}: Should load the dataset into mlr_report$task
and is in charge of labeling column roles (i.e. feature, target, group, weight)

\item \code{process(mlr_report)}: Do the training and/or run the model. Subclasses
should define a \code{train.mlr_model} and \code{predict.mlr_model} function, which should load from
\code{task} and store their reults in entries of the same name

\item \code{output(mlr_report)}: Analyze the input set and the sensitivity of the training using
some standard heuristics. TBD

\item \code{write(mlr_report)}: Save the trained model and model output for future use. Writes the model to the
\link[tessilake:cache_path]{tessilake::cache_primary_path} under the subdirectory \code{subdir} and then syncs them across storages.

}}
\keyword{datasets}
