% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2_orphans_report.R
\name{p2_orphans_report}
\alias{p2_orphans_report}
\alias{run.p2_orphans_report}
\title{p2_orphans_report}
\usage{
\method{run}{p2_orphans_report}(p2_orphans_report, freshness = 0, ...)
}
\arguments{
\item{p2_orphans_report}{p2_orphans_report object}

\item{freshness}{the returned data will be at least this fresh}

\item{...}{not used}
}
\description{
Sends an email containing a plot of recent orphans and a spreadsheet of all orphans.
}
