% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iml.R
\name{parse_shapley}
\alias{parse_shapley}
\title{parse_shapley}
\usage{
parse_shapley(explanation, filter = "=0|=NA", n = 3)
}
\arguments{
\item{explanation}{\link{data.frame} of explanations as returned by \link[iml:Shapley]{iml::Shapley}}

\item{filter}{\code{character(1)} regular expression of feature/value combinations to exclude}

\item{n}{\code{integer(1)} maximum number of explanatory features to return}
}
\value{
string explaining the feature values
}
\description{
Parse the data returned from \link[iml:Shapley]{iml::Shapley} into a single single. Explanations are sorted by importance, filtered
by \code{filter}, the top \code{n} are collected and formatted into a single string.
}
