% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf_plot}
\alias{pdf_plot}
\title{pdf_plot}
\usage{
pdf_plot(plot, title = NULL, subtitle = NULL)
}
\arguments{
\item{plot}{output from ggplot or other plot that can be \code{print}ed}

\item{title}{character optional title of the plot}

\item{subtitle}{character subtitle of the plot}
}
\description{
Add a plot to a pdf by generating R markdown, intended to be used within the \code{expr} argument to \link{write_pdf}
}
\examples{
library(ggplot2)
p <- ggplot(data.frame(x=runif(100))) + geom_histogram(aes(x))
write_pdf(pdf_plot(p, title = "Sample plot", subtitle = "histogram of random numbers"))

}
