% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf_table}
\alias{pdf_table}
\title{pdf_table}
\usage{
pdf_table(x, ...)
}
\arguments{
\item{x}{data frame to export}

\item{...}{
  Arguments passed on to \code{\link[knitr:kable]{knitr::kable}}
  \describe{
    \item{\code{digits}}{Maximum number of digits for numeric columns, passed to
\code{round()}. This can also be a vector of length \code{ncol(x)}, to set
the number of digits for individual columns.}
    \item{\code{row.names}}{Logical: whether to include row names. By default, row names
are included if \code{rownames(x)} is neither \code{NULL} nor identical to
\code{1:nrow(x)}.}
    \item{\code{col.names}}{A character vector of column names to be used in the table.}
    \item{\code{align}}{Column alignment: a character vector consisting of \code{'l'}
(left), \code{'c'} (center) and/or \code{'r'} (right). By default or if
\code{align = NULL}, numeric columns are right-aligned, and other columns
are left-aligned. If \code{length(align) == 1L}, the string will be
expanded to a vector of individual letters, e.g. \code{'clc'} becomes
\code{c('c', 'l', 'c')}, unless the output format is LaTeX.}
    \item{\code{caption}}{The table caption. By default, it is retrieved from the chunk
option \code{tab.cap}.}
    \item{\code{label}}{The table reference label. By default, the label is obtained
from \code{knitr::\link[knitr]{opts_current}$get('label')} (i.e., the current
chunk label). To disable the label, use \code{label = NA}.}
    \item{\code{format.args}}{A list of arguments to be passed to \code{\link{format}()}
to format table values, e.g. \code{list(big.mark = ',')}.}
  }}
}
\description{
Add a table to a pdf by generating R markdown, intended to be used  within the \code{expr} argument to \link{write_pdf}.
Thin wrapper around \link[knitr:kable]{knitr::kable} that generates latex output.
}
