% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_map_report.R
\docType{data}
\name{performance_map_report}
\alias{performance_map_report}
\alias{read.performance_map_report}
\alias{process.performance_map_report}
\title{performance_map_report}
\format{
An object of class \code{performance_map} (inherits from \code{report}, \code{list}) of length 0.
}
\usage{
performance_map_report

\method{read}{performance_map_report}(report)

\method{process}{performance_map_report}(report)
}
\description{
Hierarchical clustering of performances, using audience crossover
data to calculate the effective "distance" between performances
}
\section{Methods (by generic)}{
\itemize{
\item \code{read(performance_map_report)}: read data for performance map

\item \code{process(performance_map_report)}: analyze and cluster performances

}}
\keyword{datasets}
