% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetCleanR.R
\name{process.streetCleanR}
\alias{process.streetCleanR}
\title{process.streetCleanR}
\usage{
\method{process}{streetCleanR}(self)
}
\arguments{
\item{addressStream}{data.table of addresses generated by tessiStreamR::addressStream}
}
\value{
a data.table with additional columns:
\itemize{
\item suggestion.street1
\item suggestion.street2
\item suggestion.city
\item suggestion.state
\item suggestion.postal_code
\item suggestion.source
\item error (a list column of non-matching fields between the suggested and actual address)
suggestions are taken, in priority order, from Census (when matching exactly);
Google (when successful); with a fallback to Census (when matching partially).
}
}
\description{
process.streetCleanR
}
