% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_report.R
\name{report}
\alias{report}
\alias{\%+\%}
\alias{is_report}
\alias{read}
\alias{read.report}
\alias{process}
\alias{process.report}
\alias{print.report}
\alias{output}
\alias{output.report}
\alias{write}
\alias{write.report}
\alias{run}
\alias{run.report}
\title{report}
\usage{
report(x = list(), class = character())

report_x \%+\% report_y

is_report(x)

read(...)

\method{read}{report}(x, ...)

process(...)

\method{process}{report}(x, ...)

\method{print}{report}(x, ...)

output(...)

\method{output}{report}(x, ...)

write(...)

\method{write}{report}(x, ...)

run(...)

\method{run}{report}(x, ...)
}
\arguments{
\item{x}{a list}

\item{class}{an optional additional subclass}

\item{report_x}{report}

\item{report_y}{report}

\item{...}{additional parameters for subclassed methods}
}
\value{
report that contains data from \code{report_x} and \code{report_y} and inherits from both \code{report_x} and \code{report_y}
}
\description{
Base S3 class for report.
}
\section{Methods (by generic)}{
\itemize{
\item \code{read(report)}: does thing but print a warning, intended to be overloaded by subclass

\item \code{process(report)}: does thing but print a warning, intended to be overloaded by subclass

\item \code{print(report)}: prints the names and contents of the report object

\item \code{output(report)}: does thing but print a warning, intended to be overloaded by subclass

\item \code{write(report)}: does thing but print a warning, intended to be overloaded by subclass

\item \code{run(report)}: run all methods in order: read -> process -> write -> output

}}
\section{Functions}{
\itemize{
\item \code{report_x \%+\% report_y}: composable argument for report classes

\item \code{is_report()}: tests if x is an report object

\item \code{read()}: generic function to dispatch for data reading

\item \code{process()}: generic function to dispatch for data processing

\item \code{output()}: generic fucntion to dispatch for data output

\item \code{write()}: generic function to dispatch for data saving

\item \code{run()}: generic function to dispatch for report running

}}
