% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senders.R
\name{send_xlsx}
\alias{send_xlsx}
\alias{send_file}
\alias{output.email_report}
\title{send_file}
\usage{
send_xlsx(
  table,
  subject = paste(format(substitute(table)), Sys.Date()),
  body = paste("Sent by", Sys.info()["nodename"]),
  emails = config::get("tessiflow.email"),
  basename = format(substitute(table)),
  ...
)

\method{output}{email_report}(
  filename,
  subject = paste(format(substitute(table)), Sys.Date()),
  body = paste("Sent by", Sys.info()["nodename"]),
  emails = config::get("tessiflow.email"),
  basename = basename(filename),
  ...
)
}
\arguments{
\item{table}{data.table to send}

\item{subject}{character subject of the email}

\item{body}{html of message or a list containing \link[sendmailR:mime_part]{sendmailR::mime_part} objects.}

\item{emails}{email addresses (first will be sender)}

\item{basename}{name of the file to use in the email. Defaults to the name of the table.
A timestamp and extension will be appended and passed on to \link{send_email} as the name of \code{attachments}.}

\item{...}{
  Arguments passed on to \code{\link[=write_xlsx]{write_xlsx}}
  \describe{
    \item{\code{group}}{character name of column to be used for grouping/banding the data. Band color will change when column value changes.}
    \item{\code{currency}}{character vector of column to be used for grouping/banding the data. Band color will change when column value changes.}
  }}

\item{filename}{filename to write}
}
\description{
Helpers for sending a file via email
}
\section{Functions}{
\itemize{
\item \code{send_xlsx()}: Simple wrapper for \link{send_email} and \link{write_xlsx}

\item \code{output(email_report)}: report definition for sending a file by email

}}
