% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsubscribes.R
\name{send_unsubscribe_report_table}
\alias{send_unsubscribe_report_table}
\title{send_unsubscribe_report_table}
\usage{
send_unsubscribe_report_table(table, email, dry_run = FALSE)
}
\arguments{
\item{table}{data.table to send}

\item{email}{character email address to send the email to}

\item{dry_run}{boolean do not send the email if TRUE}
}
\description{
send_unsubscribe_report_table
}
