% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_report.R
\docType{data}
\name{sql_report}
\alias{sql_report}
\alias{read.sql_report}
\title{sql_report}
\format{
An object of class \code{sql_report} (inherits from \code{report}, \code{list}) of length 0.
}
\usage{
sql_report

\method{read}{sql_report}(query)
}
\arguments{
\item{report}{sql_report object}

\item{...}{not used}
}
\description{
Report that wraps the functionality of \link[tessilake:read_sql]{tessilake::read_sql} and \link{send_xlsx}
}
\section{Methods (by generic)}{
\itemize{
\item \code{read(sql_report)}: reads sql_report data
Loads data from sql
\itemize{
\item p2: unsubscribes and hard bounces by list (uses p2_stream_enriched from tessistream)
\item tessi: emails, addresses, logins, memberships, constituencies, MGOs (attributes)
}

}}
\keyword{datasets}
