% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions_model.R
\name{stream_normalize_timestamps}
\alias{stream_normalize_timestamps}
\title{stream_normalize_timestamps}
\usage{
stream_normalize_timestamps(
  dataset,
  columns = grep("timestamp", colnames(dataset), value = T),
  by = "group_customer_no"
)
}
\arguments{
\item{dataset}{data.table of data to normalize}

\item{columns}{character vector of columns to normalize; defaults to all columns with a name containing the word \code{timestamp}}

\item{by}{character column name to group the table by}
}
\value{
normalized data.table
}
\description{
Replaces the date-times in \code{columns} with integer offsets from the first \code{timestamp} per group identified by \code{by}.
}
