% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetCleanR.R
\name{streetCleanR}
\alias{streetCleanR}
\title{initialize}
\description{
initialize

initialize
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{streetCleanR$new()}}
\item \href{#method-process}{\code{streetCleanR$process()}}
\item \href{#method-report}{\code{streetCleanR$report()}}
\item \href{#method-clone}{\code{streetCleanR$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{streetCleanR$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self

auditAddressData
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process"></a>}}
\if{latex}{\out{\hypertarget{method-process}{}}}
\subsection{Method \code{process()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{streetCleanR$process(addressStream = self$data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{addressStream}}{data.table of addresses generated by tessiStreamR::addressStream}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data.table with additional columns:
\itemize{
\item suggestion.street1
\item suggestion.street2
\item suggestion.city
\item suggestion.state
\item suggestion.postal_code
\item suggestion.source
\item error (a list column of non-matching fields between the suggested and actual address)
suggestions are taken, in priority order, from Census (when matching exactly);
Google (when successful); with a fallback to Census (when matching partially).
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-report"></a>}}
\if{latex}{\out{\hypertarget{method-report}{}}}
\subsection{Method \code{report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{streetCleanR$report(addressStream = self$data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{streetCleanR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
