% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessireport-package.R
\docType{package}
\name{tessireport-package}
\alias{tessireport}
\alias{tessireport-package}
\title{tessireport: Data hygiene tools for Tessitura}
\description{
tessireport is a set of scripts for data hygiene in Tessitura, using data queried and cached by tessilake and processed by tessistream.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brooklynacademyofmusic/tessireport}
  \item Report bugs at \url{https://github.com/brooklynacademyofmusic/tessireport/issues}
}

}
\author{
\strong{Maintainer}: Sky Syzygy \email{sky@skysyzygy.xyz}

}
\keyword{internal}
