% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{write_pdf}
\alias{write_pdf}
\title{write_pdf}
\usage{
write_pdf(
  expr,
  output_file = tempfile(fileext = ".pdf"),
  .title = NULL,
  .author = NULL,
  .date = today(),
  .classoption = "landscape",
  .papersize = "legal",
  .mainfont = "Arial",
  .geometry = "margin=0.5in",
  .fontsize = "8pt",
  fig.width = 14,
  fig.height = 7.5,
  preamble = system.file("pdf_memoir-preamble.Rmd", package = "tessireport")
)
}
\arguments{
\item{expr}{expression to execute to create the body of the pdf}

\item{.title}{character title for the header of the pdf; see the preamble for how this variable is used}

\item{.author}{character author for the header of the pdf; see the preamble for how this variable is used}

\item{.date}{character date for the header of the pdf; see the preamble for how this variable is used}

\item{.classoption}{character tex classoption; see the preamble for how this variable is used}

\item{.papersize}{character tex pagesize; see the preamble for how this variable is used}

\item{.mainfont}{character tex mainfont; see the preamble for how this variable is used}

\item{.geometry}{character tex geometry; see the preamble for how this variable is used}

\item{.fontsize}{character tex fontsize; see the preamble for how this variable is used}

\item{fig.width}{character knitr figure width; see the preamble for how this variable is used}

\item{fig.height}{character knitr figure height; see the preamble for how this variable is used}

\item{preamble}{filename of the yaml preamble used by pandoc for generating the pdf}
}
\description{
write_pdf
}
