% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attendance_report.R
\docType{data}
\name{attendance_report}
\alias{attendance_report}
\alias{read.attendance_report}
\alias{process.attendance_report}
\alias{write.attendance_report}
\title{attendance_report}
\format{
An object of class \code{attendance_report} (inherits from \code{email_report}, \code{report}, \code{list}) of length 0.
}
\usage{
attendance_report

\method{read}{attendance_report}(
  attendance_report,
  list_no,
  tables = c("order_detail", "special_activities", "performances", "customers", "seats"),
  since = Sys.time(),
  until = as.POSIXct(Sys.Date() + 3),
  refresh = TRUE,
  ...
)

\method{process}{attendance_report}(
  attendance_report,
  filter = sli_status_desc \%in\% c("Unseated, Unpaid", "Seated, Unpaid", "Seated, Paid",
    "Unseated, Paid", "Upgraded", "Ticketed, Paid") | status_desc \%in\% c("Attended",
    "Accepted", "Tentative"),
  formats = c("\%b \%d", "\%b \%d \%I:\%M \%p"),
  append = NULL,
  ...
)

\method{write}{attendance_report}(
  attendance_report,
  columns = list(`customer #` = group_customer_no, `order #` =
    coalesce(as.character(order_no), ""), name = name, performance = perf_desc, time =
    perf_dt, status = status, `ship method` = ship_method, seats = seats),
  column_widths = c(0.75, 0.75, 1.75, 1, 1, 1, 1, 1),
  ...
)
}
\arguments{
\item{attendance_report}{}

\item{list_no}{\code{integer(1)} list number to load customers from}

\item{tables}{\code{character} vector of tables to load from, as defined by \link[tessilake:tessi_list_tables]{tessilake::tessi_list_tables}}

\item{since}{\code{POSIXct} date from which to query performances}

\item{until}{\code{POSIXct} date up to which to query performances}

\item{refresh}{\code{logical} whether or not to force refresh the data (default is to refresh)}

\item{...}{additional parameters passed on to \link{pdf_table}}

\item{filter}{\code{expression} to filter the output by, as with \link[rlang:eval_tidy]{rlang::eval_tidy}}

\item{formats}{\code{character(2)} vector of date and date-time formats}

\item{append}{\code{list} of additional data to append to the output on shared columns}

\item{columns}{\code{list} of named expressions mapping column names to their values,
will be evaluated in the environment of \code{attendance_report$output}}

\item{column_widths}{\code{double} vector of column widths in inches; must be the same
length as \code{columns}}
}
\description{
Sends an email containing a spreadsheet and pdf of attendees
}
\section{Methods (by generic)}{
\itemize{
\item \code{read(attendance_report)}: load data for \code{attendance_report}

\item \code{process(attendance_report)}: process data for \code{attendance_report}

\item \code{write(attendance_report)}: output a pdf of attendance

}}
\note{
\itemize{
\item \strong{order_detail} : order_no, sli_status_desc, seating_status_desc, perf_no,
perf_dt, perf_desc, recipient_display_name, customer_no, order_ship_method_desc
\item \strong{activities} : customer_no, activity_desc, status_desc
\item \strong{performances} : perf_no, perf_desc, perf_dt
\item \strong{customers} : customer_no, display_name, sort_name
\item \strong{performance_seating} : perf_no, seat_no, seat_row, seat_num
\item \strong{scans} : customer_no, perf_no
}
}
\keyword{datasets}
