% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_rollback_event}
\alias{dataset_rollback_event}
\title{dataset_rollback_event}
\usage{
dataset_rollback_event(
  dataset,
  event = "event",
  columns = NULL,
  by = "group_customer_no"
)
}
\arguments{
\item{dataset}{data.table of data to roll back}

\item{event}{character column name containing a logical feature that indicates events to rollback}

\item{columns}{character vector of columns to roll back}

\item{by}{character column name to group the table by}
}
\value{
rolled back data.table
}
\description{
Rolls back the data in \code{columns} for rows flagged by \code{event} to prevent data leaks during training.
}
