% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3proba_shim.R
\name{.__TaskSurv__initialize}
\alias{.__TaskSurv__initialize}
\title{.__TaskSurv__initialize}
\usage{
.__TaskSurv__initialize(
  id,
  backend,
  time = "time",
  event = "event",
  time2,
  type = c("right", "left", "interval", "counting", "interval2", "mstate"),
  label = NA_character_
)
}
\description{
Shim for .__TaskSurv__initialize to work with duckdb tables
}
