% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2_orphans_report.R
\docType{data}
\name{p2_orphans_report}
\alias{p2_orphans_report}
\alias{run.p2_orphans_report}
\title{p2_orphans_report}
\format{
An object of class \code{p2_orphans_report} (inherits from \code{report}, \code{list}) of length 0.
}
\usage{
p2_orphans_report

\method{run}{p2_orphans_report}(report, freshness = 0, ...)
}
\arguments{
\item{report}{sql_report object}

\item{freshness}{difftime,    the returned data will be at least this fresh}

\item{...}{not used}
}
\description{
Sends an email containing a plot of recent orphans and a spreadsheet of all orphans.
}
\section{Methods (by generic)}{
\itemize{
\item \code{run(p2_orphans_report)}: runs the p2_orphans_report

}}
\keyword{datasets}
