% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_map_report.R
\docType{data}
\name{performance_map_report}
\alias{performance_map_report}
\alias{read.performance_map}
\alias{process.performance_map}
\alias{write.performance_map}
\title{performance_map_report}
\format{
An object of class \code{performance_map} (inherits from \code{email_report}, \code{report}, \code{list}) of length 0.
}
\usage{
performance_map_report

\method{read}{performance_map}(
  report,
  since = Sys.Date() - 365 * 5,
  until = Sys.Date() + 365,
  filter_expr = NULL,
  ...
)

\method{process}{performance_map}(report, n_clusters = 8, ...)

\method{write}{performance_map}(report, n_clusters = 8, highlight_since = Sys.Date() - 365, ...)
}
\arguments{
\item{report}{\code{report} object}

\item{since}{\code{POSIXct} performance data on/after this date will be returned}

\item{until}{\code{POSIXct} performance data on/before this date will be returned}

\item{filter_expr}{\code{expression} to use when filtering performances, i.e.
\code{read_tessi("performances") \%>\% filter(!!filter_expr)}}

\item{...}{not used}

\item{n_clusters}{\code{integer} number of clusters to make}

\item{highlight_since}{\code{POSIXct} performances on/after this date will be highlighted in the returned plots}
}
\description{
Hierarchical clustering of performances, using audience crossover
data to calculate the effective "distance" between performances.
\itemize{
\item Distance is calculated using Simpson's distance metric
\item Distance is corrected by time using a binomial regression on time between
performances
\item Clustering is hierarchical using Ward D2 distance
\item Mapping is classic multidimensional scaling, i.e. PCA
}
}
\section{Functions}{
\itemize{
\item \code{read(performance_map)}: read data for performance map

\item \code{process(performance_map)}: analyze and cluster performances

\item \code{write(performance_map)}: write pdf of performance maps and an file for importing into Tessitura

}}
\keyword{datasets}
