% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditR.R
\name{report}
\alias{report}
\alias{new_report}
\alias{is_report}
\alias{read}
\alias{process}
\alias{print.report}
\alias{output}
\alias{write}
\title{report}
\usage{
new_report(x = new.env(), class = character())

is_report(x)

read(x, ...)

process(x, ...)

\method{print}{report}(x, ...)

output(x, ...)

write(x, ...)
}
\arguments{
\item{x}{an environment}

\item{class}{an optional additional subclass}
}
\description{
Base S3 class for report.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(report)}: prints the names and contents of the report object

}}
\section{Functions}{
\itemize{
\item \code{is_report()}: tests if x is an report object

\item \code{read()}: generic function to dispatch for data reading

\item \code{process()}: generic function to dispatch for data processing

\item \code{output()}: generic fucntion to dispatch for data output

\item \code{write()}: generic function to dispatch for data saving

}}
