% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senders.R
\name{send_email}
\alias{send_email}
\title{send_email}
\usage{
send_email(
  subject,
  body,
  emails = config::get("tessiflow.email"),
  smtp = config::get("tessiflow.smtp"),
  ...
)
}
\arguments{
\item{subject}{email subject}

\item{body}{email body}

\item{emails}{email addresses (first will be sender)}

\item{smtp}{smtp configuration}

\item{...}{additional parameters sent on to \link[mailR:send.mail]{mailR::send.mail}}
}
\description{
send_email
}
