% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senders.R
\name{send_file}
\alias{send_file}
\alias{send_xlsx}
\alias{output.email_report}
\title{send_file}
\usage{
send_file(
  filename,
  name = basename(filename),
  subject = paste(name, Sys.Date()),
  ...
)

send_xlsx(table, basename = format(substitute(table)), ...)

\method{output}{email_report}(report, report_filename = "filename", ...)
}
\arguments{
\item{filename}{filename to write}

\item{subject}{character subject of the email}

\item{...}{
  Arguments passed on to \code{\link[=send_email]{send_email}}, \code{\link[=write_xlsx]{write_xlsx}}, \code{\link[=send_email]{send_email}}, \code{\link[=send_email]{send_email}}
  \describe{
    \item{\code{emails}}{email addresses (first will be sender)}
    \item{\code{smtp}}{named list, should contain \code{hostname} and (optionally) \code{port}, which defaults to 25.
Additional elements will get merged with \code{...} and passed to curl, see \link[curl:curl_options]{curl::curl_options} for more details.}
    \item{\code{body}}{html of message or a list containing \link[sendmailR:mime_part]{sendmailR::mime_part} objects.}
    \item{\code{engine}}{One of: \itemize{
\item{\code{"internal"} for the internal smtp transport (default).}
\item{\code{"curl"} for the use of curl for transport. Enable if you need STARTTLS/SSL
and/or SMTP authentication. See \code{curl::\link[curl]{send_mail}}.}
\item{\code{"debug"} sendmail returns a RFC2822 formatted email message without sending it.}
}}
    \item{\code{group}}{character name of column to be used for grouping/banding the data. Band color will change when column value changes.}
    \item{\code{currency}}{character vector of column to be used for grouping/banding the data. Band color will change when column value changes.}
  }}

\item{table}{data.table to send}

\item{basename}{name of the file to use in the email. Defaults to the name of the table or the inputted filename.
A timestamp and extension will be appended and passed on to \link{send_email} as the name of \code{attachments}.}

\item{report}{\code{report} object}

\item{report_filename}{\code{character(1)} name of the field that contains the filename}
}
\description{
Helpers for sending a file via email
}
\section{Functions}{
\itemize{
\item \code{send_xlsx()}: Simple wrapper for \link{send_file} and \link{write_xlsx}

\item \code{output(email_report)}: report definition for sending a file by email

}}
