% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsubscribe_report.R
\name{send_unsubscribe_report_table}
\alias{send_unsubscribe_report_table}
\title{send_unsubscribe_report_table}
\usage{
send_unsubscribe_report_table(table, email)
}
\arguments{
\item{table}{data.table to send}

\item{email}{character email address to send the email to}
}
\description{
Simple wrapper for \link{send_xlsx}
}
