% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senders.R
\name{send_xlsx}
\alias{send_xlsx}
\title{send_xlsx}
\usage{
send_xlsx(
  table,
  subject = paste(substitute(table), Sys.Date()),
  body = paste("Sent by", Sys.info()["nodename"]),
  emails = config::get("tessiflow.email"),
  dry_run = FALSE,
  ...
)
}
\arguments{
\item{table}{data.table to send}

\item{subject}{character subject of the email, default is the name of the \code{table} and the current date.}

\item{body}{character body of the email, default is a human readable message indicating the computer name.}

\item{dry_run}{boolean do not send the email if TRUE}

\item{...}{additional parameters to send on to \link{send_email} and then to \link[mailR:send.mail]{mailR::send.mail}}

\item{email}{character email addresses to send the email to (first will be sender as well)}
}
\description{
Simple wrapper for \link{send_email} and \link{write_xlsx}
}
\note{
Useful parameters are \code{subject}, \code{body}, \code{emails}, \code{html}, \code{inline}, and \code{file.names}
}
