% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senders.R
\name{send_xlsx}
\alias{send_xlsx}
\title{send_xlsx}
\usage{
send_xlsx(
  table,
  subject = paste(format(substitute(table)), Sys.Date()),
  body = paste("Sent by", Sys.info()["nodename"]),
  emails = config::get("tessiflow.email"),
  basename = format(substitute(table)),
  ...
)
}
\arguments{
\item{table}{data.table to send}

\item{subject}{Subject of the email.}

\item{body}{Body of the email as text. If the parameter body refers to an existing file location,
the text of the file is parsed as body of the email. Default is \verb{Sent by <computer name>}.}

\item{emails}{email addresses (first will be sender)}

\item{basename}{name of the file to use in the email. Defaults to the name of the table.
A timestamp and extension will be appended and passed on to \code{mailR::send.mail} as \code{file.names}}

\item{...}{
  Arguments passed on to \code{\link[mailR:send.mail]{mailR::send.mail}}, \code{\link[=write_xlsx]{write_xlsx}}
  \describe{
    \item{\code{html}}{A boolean indicating whether the body of the email should be parsed as HTML.}
    \item{\code{inline}}{A boolean indicating whether images in the HTML file should be embedded inline.}
    \item{\code{group}}{character name of column to be used for grouping/banding the data. Band color will change when column value changes.}
    \item{\code{currency}}{character vector of column to be used for grouping/banding the data. Band color will change when column value changes.}
  }}
}
\description{
Simple wrapper for \link{send_email} and \link{write_xlsx}
}
