% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senders.R
\name{send_xlsx}
\alias{send_xlsx}
\title{send_xlsx}
\usage{
send_xlsx(
  table,
  subject = paste(format(substitute(table)), Sys.Date()),
  body = paste("Sent by", Sys.info()["nodename"]),
  emails = config::get("tessiflow.email"),
  ...
)
}
\arguments{
\item{table}{data.table to send}

\item{subject}{Subject of the email.}

\item{body}{Body of the email as text. If the parameter body refers to an existing file location, the text of the file is parsed as body of the email.}

\item{emails}{email addresses (first will be sender)}

\item{...}{
  Arguments passed on to \code{\link[mailR:send.mail]{mailR::send.mail}}
  \describe{
    \item{\code{html}}{A boolean indicating whether the body of the email should be parsed as HTML.}
    \item{\code{inline}}{A boolean indicating whether images in the HTML file should be embedded inline.}
  }}
}
\description{
Simple wrapper for \link{send_email} and \link{write_xlsx}
}
