% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetCleanR.R
\name{new_streetCleanR}
\alias{new_streetCleanR}
\title{streetCleanR}
\usage{
new_streetCleanR(x=new.env())
}
\arguments{
\item{x}{an environment of data}
}
\description{
Address auditing based on addressStream
}
