% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tessireport-package.R
\docType{package}
\name{tessireport-package}
\alias{tessireport}
\alias{tessireport-package}
\title{tessireport: Data hygiene tools for Tessitura}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/skysyzygy/tessireport}
  \item Report bugs at \url{https://github.com/skysyzygy/tessireport/issues}
}

}
\author{
\strong{Maintainer}: Sky Syzygy \email{sky@skysyzygy.xyz} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
