% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsubscribes.R
\name{unsubscribe_report}
\alias{unsubscribe_report}
\alias{new_unsubscribe_report}
\alias{read.unsubscribe_report}
\alias{process.unsubscribe_report}
\title{unsubscribe_report}
\usage{
new_unsubscribe_report()

\method{read}{unsubscribe_report}(report, customers)

\method{process}{unsubscribe_report}(report)
}
\arguments{
\item{report}{report object}

\item{customers}{integer vector of customer numbers to load}
}
\description{
Reports on unsubscribes and bad addresses for a group of customers
}
\section{Methods (by generic)}{
\itemize{
\item \code{read(unsubscribe_report)}: reads unsubscribe_report data

\item \code{process(unsubscribe_report)}: process unsubscribe_report data

}}
\section{Functions}{
\itemize{
\item \code{new_unsubscribe_report()}: creates a new unsubscribe_report object

}}
