% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsubscribes.R
\name{unsubscribe_report}
\alias{unsubscribe_report}
\alias{read.unsubscribe_report}
\alias{process.unsubscribe_report}
\alias{output.unsubscribe_report}
\title{unsubscribe_report}
\usage{
unsubscribe_report()

\method{read}{unsubscribe_report}(report, customers, ...)

\method{process}{unsubscribe_report}(report, ...)

\method{output}{unsubscribe_report}(report, since = Sys.Date() - 30, until = Sys.Date() + 30, ...)
}
\arguments{
\item{report}{report object}

\item{customers}{integer vector of customer numbers to load}

\item{...}{not used}
}
\description{
Reports on unsubscribes and bad addresses for a group of customers
}
\section{Methods (by generic)}{
\itemize{
\item \code{read(unsubscribe_report)}: reads unsubscribe_report data
Loads data from
\itemize{
\item p2: unsubscribes and hard bounces by list (uses p2_stream_enriched from tessistream)
\item tessi: emails, addresses, logins, memberships, constituencies, MGOs (attributes)
}

\item \code{process(unsubscribe_report)}: process unsubscribe_report data
\itemize{
\item Checks if emails are unsubscribed, bounced, or don't exist
\item Checks if mailing addresses have a NCOA flag or don't exist
\item Checks if the customer primary login doesn't match the primary email
\item Checks if customers are inactive
\item Adds identifying info: MGO, constituencies, membership expiration date and level
}

\item \code{output(unsubscribe_report)}: send the unsubscribe_report as emails to MGOs
Routes based on the following rules:
\itemize{
\item MGO/PSD signatory -> send to users with matching name
\item GOV -> send to Government Affairs (cluna)
\item CP# -> send to Strategic Partnerships (ashah)
\item Patron -> send to Patron Program (jhindle)
\item Other -> send to Dev Ops (kburke)
}

}}
\section{Functions}{
\itemize{
\item \code{unsubscribe_report()}: creates a new unsubscribe_report object

}}
