% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsubscribes.R
\docType{data}
\name{unsubscribe_report}
\alias{unsubscribe_report}
\alias{read.unsubscribe_report}
\alias{process.unsubscribe_report}
\alias{output.unsubscribe_report}
\title{unsubscribe_report}
\format{
An object of class \code{unsubscribe_report} (inherits from \code{report}, \code{list}) of length 0.
}
\usage{
unsubscribe_report

\method{read}{unsubscribe_report}(report, customers, ...)

\method{process}{unsubscribe_report}(report, ...)

\method{output}{unsubscribe_report}(
  report,
  since = Sys.Date() - 30,
  until = Sys.Date() + 30,
  routing_rules = list(TRUE ~ list(config::get("tessiflow.email"))),
  ...
)
}
\arguments{
\item{report}{report object}

\item{customers}{integer vector of customer numbers to load}

\item{...}{not used}

\item{since}{date, start date for filtering the returned data}

\item{until}{date, end date for filtering the returned data}

\item{routing_rules}{list of formulas to be used for routing, as for \link[dplyr:case_when]{dplyr::case_when}. Can refer to any columns returned by \link{process.unsubscribe_report}. See \strong{Note} below}
}
\description{
Reports on unsubscribes and bad addresses for a group of customers
}
\section{Methods (by generic)}{
\itemize{
\item \code{read(unsubscribe_report)}: reads unsubscribe_report data
Loads data from
\itemize{
\item p2: unsubscribes and hard bounces by list (uses p2_stream_enriched from tessistream)
\item tessi: emails, addresses, logins, memberships, constituencies, MGOs (attributes)
}

\item \code{process(unsubscribe_report)}: process unsubscribe_report data
\itemize{
\item Checks if emails are unsubscribed, bounced, or don't exist
\item Checks if mailing addresses have a NCOA flag or don't exist
\item Checks if the customer primary login doesn't match the primary email
\item Checks if customers are inactive
\item Adds identifying info: MGO, constituencies, membership expiration date and level
}

\item \code{output(unsubscribe_report)}: send the unsubscribe_report as emails to MGOs

Data is filtered so that all events after \code{since} are returned and all events for members with expiration dates
between \code{since} and \code{until} are returned

}}
\section{Functions}{
\itemize{
\item \code{unsubscribe_report}: creates a new unsubscribe_report object

}}
\note{
For example, this value for routing_rules:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(grepl("GOV", constituencies) ~ list("eleszynski@bam.org"),
     grepl("CP\\\\d", constituencies) ~ list("lmcgee@bam.org"),
     grepl("\\\\+",constituencies) ~ list(c("apratama@bam.org","jhindle@bam.org")),
     TRUE ~ list(c("kburke@bam.org","esearles@bam.org")))
}\if{html}{\out{</div>}}

Routes based on the following rules:
\itemize{
\item GOV -> send to Government Affairs (eleszynski)
\item CP# -> send to Strategic Partnerships (lmcgee)
\item Patron -> send to Patron Program (apratama and jhindle)
\item Other -> send to Dev Ops (kburke and esearles)
}
}
\keyword{datasets}
