% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{write_xlsx}
\usage{
write_xlsx(data, filename = tempfile("write_xlsx", fileext = ".xlsx"))
}
\arguments{
\item{data}{data to write to the spreadsheet}

\item{filename}{filename to write}
}
\value{
filename of written xlsx
}
\description{
Convenience wrapper around \code{openxlsx::write.xlsx} to handle some common formatting
\itemize{
\item column headers are title-cased and bolded
\item column widths are calculated based on the string representations of the data
\item Date columns are formatted as DATE instead of LONGDATE
}
}
